Fix install pipeline, media viewer recreation, cache leak, headers, and 276 compile errors
Fixes #1.

Planned fixes:
- InstallSplitApkFilesJob now groups cached APKs by package name so
  unrelated apps go through separate PackageInstaller sessions while a
  same-app base+split set stays in one atomic session.
- PackageInstallerStatusReceiver routes the pending-user-action confirm
  intent through BackgroundActivityStarter: direct start when foreground,
  a high-priority notification tap when backgrounded (Android 10+ drops
  plain startActivity from receivers, leaving sessions stuck).
- FileJobService.onCreate sweeps stale split-bundle-*/split-file-* APKs
  from cacheDir; previously only the job's finally block deleted them.
- Media viewer keeps ExoPlayer in a new MediaViewerViewModel
  (released in onCleared) so position/buffering survive recreation;
  fixed the stale KDoc claiming retainInstance on an Activity.
- Restored GPL header blocks on 5 files where the dual-pane work had
  stripped them; added headers to 3 new files (receiver + 2 layouts).

Pre-existing compile errors (276 -> 0, compileDebugKotlin now green):
- Kotlin nested block comments broke whole files (META-INF/* inside
  KDoc) in ApkSignatureInfo, ApkV1Signer, DavResourceCompat.
- dav4jvm 2.1.4: property.webdav subpackage removed -> property.*;
  ResponseCallback/proppatch removed; HttpUtils returns Date;
  BasicDigestAuthHandler takes a String password.
- dexlib2 2.5.2: smaliTreeWalker uses setters, Opcodes.api (no
  apiLevel/dexVersion), class extraction via classSection,
  ImmutableMethod needs ImmutableMethodParameter/Annotation conversion.
- retrofile NIO: no FileVisitOption.NOFOLLOW_LINKS (empty set),
  Paths.get needs a String/File path.
- Missing/broken imports (readAttributes, valueCompat, layoutInflater,
  copyText, Intent, duplicate ParcelableArgs), WakeWifiLock property
  API, include-binding isVisible via root, KEYCODE_MOVE_END,
  Request\$Builder escaping in HookPresets, duplicate companion object
  in CreateArchiveDialogFragment, RemoteInvoker signature, terminal
  when-branch exhaustiveness, suspendCancellableCoroutine generics.

+++ app/src/main/java/me/zhanghai/android/files/apksign/ApkV2V3Signer.kt
@@ -62,8 +62,8 @@ internal object ApkV2V3Signer {
     private const val PSS_SALT_LENGTH = 32
 
     // Block IDs for the APK Signing Block pairs.
-    private const val V2_BLOCK_ID = 0x7109871A
-    private const val V3_BLOCK_ID = 0xF05368C0
+    private const val V2_BLOCK_ID = 0x7109871A.toInt()
+    private const val V3_BLOCK_ID = 0xF05368C0.toInt()
 
     private val APK_SIGNING_BLOCK_MAGIC =
         byteArrayOf( // "APK Sig Block 42"
@@ -165,7 +165,7 @@ internal object ApkV2V3Signer {
         val rangeSize = end - start
         val sha = MessageDigest.getInstance("SHA-256")
         // Top-level: 0x5a | uint32_le(chunk_count) | concat(chunk_digests)
-        sha.update(0x5a)
+        sha.update(0x5a.toByte())
         if (rangeSize <= 0) {
             sha.update(BytesLe.uint32(0)) // chunk_count = 0
             return sha.digest()
@@ -188,7 +188,7 @@ internal object ApkV2V3Signer {
             file.seek(offset)
             file.readFully(buffer, 0, chunkLen)
             val chunkSha = MessageDigest.getInstance("SHA-256")
-            chunkSha.update(0xA5)
+            chunkSha.update(0xA5.toByte())
             chunkSha.update(BytesLe.uint32(chunkLen.toLong()))
             chunkSha.update(buffer, 0, chunkLen)
             sha.update(chunkSha.digest())
